+++ openstack/cce/v3/common_tests.go
@@ -9,6 +9,6 @@ const TokenID = client.TokenID
 
 func ServiceClient() *golangsdk.ServiceClient {
 	sc := client.ServiceClient()
-	sclient.ResourceBase = sclient.Endpoint + "api/" + "v3/" + "projects/" + "c59fd21fd2a94963b822d8985b884673/"
+	sc.ResourceBase = sc.Endpoint + "api/" + "v3/" + "projects/" + "c59fd21fd2a94963b822d8985b884673/"
 	return sc
 }
