+++ openstack/cce/v3/nodes/update.go
@@ -1,9 +1,21 @@
 package nodes
 
-import "github.com/opentelekomcloud/gophertelekomcloud"
+import (
+	"github.com/opentelekomcloud/gophertelekomcloud"
+	"github.com/opentelekomcloud/gophertelekomcloud/internal/extract"
+)
+
+// UpdateOpts contains all the values needed to update a new node
+type UpdateOpts struct {
+	Metadata UpdateMetadata `json:"metadata,omitempty"`
+}
+
+type UpdateMetadata struct {
+	Name string `json:"name,omitempty"`
+}
 
 // Update allows nodes to be updated.
-func Update(client *golangsdk.ServiceClient, clusterID, nodeID string, opts UpdateOpts) (r UpdateResult) {
+func Update(client *golangsdk.ServiceClient, clusterID, nodeID string, opts UpdateOpts) (*Nodes, error) {
 	b, err := golangsdk.BuildRequestBody(opts, "")
 	if err != nil {
 		return nil, err
@@ -12,5 +24,11 @@ func Update(client *golangsdk.ServiceClient, clusterID, nodeID string, opts Upda
 	raw, err := client.Put(client.ServiceURL("clusters", clusterID, "nodes", nodeID), b, nil, &golangsdk.RequestOpts{
 		OkCodes: []int{200},
 	})
-	return
+	if err != nil {
+		return nil, err
+	}
+
+	var res Nodes
+	err = extract.Into(raw, &res)
+	return &res, err
 }
